data(plantings): constrain positive quantities
Model validators protect API writes but bulk updates, scripts, and future code paths can bypass them. Audit every aggregate and per-cell planting table with actionable row IDs before adding minimum-one checks, so deployment fails safely instead of accepting or obscuring corrupt quantity history.
